+++ regression/docker-compose.yml
@@ -22,29 +22,25 @@
 
 version: "3.4"
 
-
-x-common: &common
-  context: ./../
-
-
-
 services:
 
-  identity-server:
+  _base_image:
     build:
-      <<: *common
       dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/identity-server
+      context: ./../
+    image: bitdust-baseimage
+    entrypoint: ""
+    command: /bin/false
+
+  identity-server:
+    image: bitdust-baseimage
     ports:
         - "22201:22"
 
 
 
   stun_1:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/stun
+    image: bitdust-baseimage
     ports:
       - "22211:22"
     links:
@@ -55,10 +51,7 @@ services:
       - "dht_seed_4"
 
   stun_2:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/stun
+    image: bitdust-baseimage
     ports:
       - "22212:22"
     links:
@@ -71,28 +64,19 @@ services:
 
 
   dht_seed_0:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/dht_seed
+    image: bitdust-baseimage
     ports:
       - "22220:22"
 
   dht_seed_1:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/dht_seed
+    image: bitdust-baseimage
     ports:
       - "22221:22"
     links:
       - "dht_seed_0"
 
   dht_seed_2:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/dht_seed
+    image: bitdust-baseimage
     ports:
       - "22222:22"
     links:
@@ -100,10 +84,7 @@ services:
       - "dht_seed_1"
 
   dht_seed_3:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/dht_seed
+    image: bitdust-baseimage
     ports:
       - "22223:22"
     links:
@@ -112,10 +93,7 @@ services:
       - "dht_seed_2"
 
   dht_seed_4:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/dht_seed
+    image: bitdust-baseimage
     ports:
       - "22224:22"
     links:
@@ -127,10 +105,7 @@ services:
 
 
   proxy_server_1:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/proxy_server
+    image: bitdust-baseimage
     ports:
       - "22231:22"
     links:
@@ -144,10 +119,7 @@ services:
       - "identity-server:is"
 
   proxy_server_2:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/proxy_server
+    image: bitdust-baseimage
     ports:
       - "22232:22"
     links:
@@ -163,10 +135,7 @@ services:
 
 
   supplier_1:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/supplier
+    image: bitdust-baseimage
     ports:
       - "22241:22"
     links:
@@ -180,10 +149,7 @@ services:
       - "identity-server:is"
 
   supplier_2:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/supplier
+    image: bitdust-baseimage
     ports:
       - "22242:22"
     links:
@@ -197,10 +163,7 @@ services:
       - "identity-server:is"
 
   supplier_3:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/supplier
+    image: bitdust-baseimage
     ports:
       - "22243:22"
     links:
@@ -214,10 +177,7 @@ services:
       - "identity-server:is"
 
   supplier_4:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/supplier
+    image: bitdust-baseimage
     ports:
       - "22244:22"
     links:
@@ -231,10 +191,7 @@ services:
       - "identity-server:is"
 
   supplier_5:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/supplier
+    image: bitdust-baseimage
     ports:
       - "22245:22"
     links:
@@ -248,10 +205,7 @@ services:
       - "identity-server:is"
 
   supplier_6:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/supplier
+    image: bitdust-baseimage
     ports:
       - "22246:22"
     links:
@@ -265,10 +219,7 @@ services:
       - "identity-server:is"
 
   supplier_7:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/supplier
+    image: bitdust-baseimage
     ports:
       - "22247:22"
     links:
@@ -282,10 +233,7 @@ services:
       - "identity-server:is"
 
   supplier_8:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/supplier
+    image: bitdust-baseimage
     ports:
       - "22248:22"
     links:
@@ -301,10 +249,7 @@ services:
 
 
   customer_1:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/customer
+    image: bitdust-baseimage
     ports:
       - "22251:22"
     links:
@@ -320,10 +265,7 @@ services:
       - customer_1:/customer_1
 
   customer_2:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/customer
+    image: bitdust-baseimage
     ports:
       - "22252:22"
     links:
@@ -339,10 +281,7 @@ services:
       - customer_2:/customer_2
 
   customer_3:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/customer
+    image: bitdust-baseimage
     ports:
       - "22253:22"
     links:
@@ -358,10 +297,7 @@ services:
       - customer_3:/customer_3
 
   customer_4:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/customer
+    image: bitdust-baseimage
     ports:
       - "22254:22"
     links:
@@ -377,10 +313,7 @@ services:
       - customer_4:/customer_4
 
   customer_5:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/customer
+    image: bitdust-baseimage
     ports:
       - "22255:22"
     links:
@@ -396,10 +329,7 @@ services:
       - customer_5:/customer_5
 
   customer_backup:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/customer
+    image: bitdust-baseimage
     ports:
       - "22256:22"
     links:
@@ -415,10 +345,7 @@ services:
       - customer_backup:/customer_backup
 
   customer_restore:
-    build:
-      <<: *common
-      dockerfile: regression/Dockerfiles/bitdust_install/Dockerfile
-    image: bitdust/customer
+    image: bitdust-baseimage
     ports:
       - "22257:22"
     links:
@@ -435,7 +362,7 @@ services:
 
   tester:
     build:
-      <<: *common
+      context: ./../
       dockerfile: regression/Dockerfiles/tester/Dockerfile
     links:
       - "identity-server:is"
