+++ regression/tests/conftest.py
@@ -23,20 +23,17 @@
 import os
 import pytest
 import time
-import requests
 import asyncio
 import pprint
-import aiohttp  # @UnresolvedImport
 
-from .testsupport import run_ssh_command_and_wait, open_tunnel, open_tunnel_async, tunnel_url, run_ssh_command_and_wait_async
 
+from . import testsupport as ts
 
 #------------------------------------------------------------------------------
 
 # DHT_SEED_NODES = 'dht_seed_0:14441, dht_seed_1:14441, dht_seed_2:14441, dht_seed_3:14441, dht_seed_4:14441, stun_1:14441, stun_2:14441'
 DHT_SEED_NODES = 'dht_seed_0:14441'
 
-PROXY_ROUTERS = 'http://is:8084/proxy_server_1.xml http://is:8084/proxy_server_2.xml'
 
 # TODO: keep this list up to date with docker-compose links
 ALL_NODES = [
@@ -111,543 +108,36 @@ ALL_ROLES = {
     ],
 }
 
-# ALL_NODES = [n[0] for r in ALL_ROLES.values() for n in r]
-
-#------------------------------------------------------------------------------
-
-def health_check(node):
-    count = 0
-    while True:
-        if count > 10:
-            assert False, f'node {node} is not healthy after 10 attempts'
-        try:
-            response = requests.get(tunnel_url(node, 'process/health/v1'))
-        except Exception as exc:
-            # print(f'[{node}] retry {count}   GET:process/health/v1  :  {exc}')
-            response = None
-        if response and response.status_code == 200 and response.json()['status'] == 'OK':
-            break
-        print(f'node {node} process not started yet, try again after 1 sec.', flush=True)
-        time.sleep(1)
-        count += 1
-    print(f'process/health/v1 [{node}] : OK\n')
-
-
-async def health_check_async(node, event_loop):
-    count = 0
-    while True:
-        if count > 10:
-            assert False, f'node {node} is not healthy after 10 attempts'
-        try:
-            client = aiohttp.ClientSession(loop=event_loop)
-            response = await client.get(tunnel_url(node, 'process/health/v1'))
-        except:
-            response = None
-        if response and response.status_code == 200 and response.json()['status'] == 'OK':
-            break
-        print(f'node {node} process not started yet, try again after 1 sec.', flush=True)
-        time.sleep(1)
-        count += 1
-    print(f'process/health/v1 [{node}] : OK\n')
-
-
-def create_identity(node, identity_name):
-    count = 0
-    while True:
-        if count > 10:
-            assert False, f'node {node} failed to create identity after 10 retries'
-        response = requests.post(
-            url=tunnel_url(node, 'identity/create/v1'),
-            json={
-                'username': identity_name,
-            },
-        )
-        if response.json()['status'] == 'OK':
-            # print('\n' + response.json()['result'][0]['xml'] + '\n')
-            break
-        if not response.status_code == 200 or (response.json()['status'] == 'ERROR' and response.json()['errors'][0] == 'network connection error'):
-            count += 1
-            # print('[%s] retry %d   POST:identity/create/v1  username=%s     network connection error' % (node, count, identity_name, ))
-            continue
-
-        assert False, f'[{node}] bad response from /identity/create/v1'
-
-    print(f'identity/create/v1 [{node}] with name {identity_name} : OK\n')
-
-
-async def create_identity_async(node, identity_name, event_loop):
-    for i in range(10):
-        client = aiohttp.ClientSession(loop=event_loop)
-        response_identity = await client.post(tunnel_url(node, 'identity/create/v1'), json={'username': identity_name})
-        assert response_identity.status == 200
-
-        response_json = await response_identity.json()
-
-        if response_json['status'] == 'OK':
-            # print('\n' + response_json['result'][0]['xml'] + '\n')
-            break
-        else:
-            assert response_json['errors'] == ['network connection error'], response_json
-
-        print('[%s] retry %d   POST:identity/create/v1  username=%s    after 1 sec.' % (node, i + 1, identity_name,))
-        await asyncio.sleep(1)
-    else:
-        assert False
-
-    print('identity/create/v1 [%s] with name %s : OK\n' % (node, identity_name,))
-
-
-def connect_network(node):
-    count = 0
-    response = requests.get(url=tunnel_url(node, 'network/connected/v1?wait_timeout=1'))
-    assert response.json()['status'] == 'ERROR'
-    while True:
-        if count > 10:
-            assert False, f'node {node} failed to connect to the network after few retries'
-        response = requests.get(tunnel_url(node, 'network/connected/v1?wait_timeout=5'))
-        if response.json()['status'] == 'OK':
-            break
-        count += 1
-        # print('[%s] retry %d   GET:network/connected/v1' % (node, count, ))
-        time.sleep(1)
-    print(f'network/connected/v1 [{node}] : OK\n')
-
-
-async def connect_network_async(node, loop):
-    client = aiohttp.ClientSession(loop=loop)
-    response = await client.get(tunnel_url(node, 'network/connected/v1?wait_timeout=1'))
-    response_json = await response.json()
-    assert response_json['status'] == 'ERROR'
-
-    for i in range(5):
-        response = await client.get(f'http://{node}:8180/network/connected/v1?wait_timeout=1')
-        response_json = await response.json()
-        if response_json['status'] == 'OK':
-            print(f"network/connected/v1 {node}: got status OK", flush=True)
-            break
-
-        print(f"{node}: sleep 1 sec", flush=True)
-        await asyncio.sleep(1)
-    else:
-        assert False
-
-#------------------------------------------------------------------------------
-
-def start_daemon(node):
-    run_ssh_command_and_wait(node, 'mkdir -pv /root/.bitdust/metadata/')
-    run_ssh_command_and_wait(node, 'echo "docker" > /root/.bitdust/metadata/networkname')
-    bitdust_daemon = run_ssh_command_and_wait(node, 'bitdust daemon')
-    print('\n' + bitdust_daemon[0].strip())
-    assert (
-        bitdust_daemon[0].strip().startswith('main BitDust process already started') or
-        bitdust_daemon[0].strip().startswith('new BitDust process will be started in daemon mode')
-    )
-    print(f'start_daemon [{node}] OK\n')
-
-
-async def start_daemon_async(node, loop):
-    await run_ssh_command_and_wait_async(node, 'mkdir -pv /root/.bitdust/metadata/', loop)
-    await run_ssh_command_and_wait_async(node, 'echo "docker" > /root/.bitdust/metadata/networkname', loop)
-    bitdust_daemon = await run_ssh_command_and_wait_async(node, 'bitdust daemon', loop)
-    print('\n' + bitdust_daemon[0].strip())
-    assert (
-        bitdust_daemon[0].strip().startswith('main BitDust process already started') or
-        bitdust_daemon[0].strip().startswith('new BitDust process will be started in daemon mode')
-    )
-    print(f'start_daemon_async [{node}] OK\n')
-
-
-def stop_daemon(node, skip_checks=False):
-    bitdust_stop = run_ssh_command_and_wait(node, 'bitdust stop')
-    print('\n' + bitdust_stop[0].strip())
-    if not skip_checks:
-        assert (
-            (
-                bitdust_stop[0].strip().startswith('BitDust child processes found') and
-                bitdust_stop[0].strip().endswith('BitDust stopped')
-            ) or (
-                bitdust_stop[0].strip().startswith('found main BitDust process:') and
-                bitdust_stop[0].strip().endswith('BitDust process finished correctly')
-            )
-        )
-    print(f'stop_daemon [{node}] OK\n')
-
-#------------------------------------------------------------------------------
-
-def start_identity_server(node):
-    print(f'\nNEW IDENTITY SERVER at [{node}]\n')
-    # use short key to run tests faster
-    run_ssh_command_and_wait(node, 'bitdust set personal/private-key-size 1024')
-    # disable unwanted services
-    run_ssh_command_and_wait(node, 'bitdust set services/customer/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/supplier/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/proxy-transport/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/proxy-server/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/private-messages/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/nodes-lookup/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/identity-propagate/enabled false')
-    # configure DHT udp port
-    run_ssh_command_and_wait(node, 'bitdust set services/entangled-dht/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/entangled-dht/udp-port "14441"')
-    # configure and enable ID server
-    run_ssh_command_and_wait(node, f'bitdust set services/identity-server/host {node}')
-    run_ssh_command_and_wait(node, 'bitdust set services/identity-server/enabled true')
-    # start BitDust daemon
-    start_daemon(node)
-    health_check(node)
-    print(f'\nSTARTED IDENTITY SERVER [{node}]\n')
-
-
-def start_dht_seed(node, wait_seconds=0, other_seeds=''):
-    print(f'\nNEW DHT SEED (with STUN SERVER) at [{node}]\n')
-    # use short key to run tests faster
-    run_ssh_command_and_wait(node, 'bitdust set personal/private-key-size 1024')
-    # disable unwanted services
-    run_ssh_command_and_wait(node, 'bitdust set services/customer/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/supplier/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/proxy-transport/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/proxy-server/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/private-messages/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/nodes-lookup/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/identity-propagate/enabled false')
-    # configure DHT udp port and seed nodes
-    run_ssh_command_and_wait(node, 'bitdust set services/entangled-dht/udp-port "14441"')
-    run_ssh_command_and_wait(node, 'bitdust set services/entangled-dht/known-nodes "%s"' % other_seeds)
-    
-    # enable Stun server
-    run_ssh_command_and_wait(node, 'bitdust set services/ip-port-responder/enabled true')
-    # start BitDust daemon
-    print(f'sleep {wait_seconds} seconds')
-    time.sleep(wait_seconds)
-    start_daemon(node)
-    health_check(node)
-    print(f'\nSTARTED DHT SEED (with STUN SERVER) [{node}]\n')
-
-
-def start_stun_server(node):
-    print(f'\nNEW STUN SERVER at [{node}]\n')
-    # use short key to run tests faster
-    run_ssh_command_and_wait(node, 'bitdust set personal/private-key-size 1024')
-    # disable unwanted services
-    run_ssh_command_and_wait(node, 'bitdust set services/customer/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/supplier/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/proxy-transport/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/proxy-server/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/private-messages/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/nodes-lookup/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/identity-propagate/enabled false')
-    # configure DHT udp port
-    run_ssh_command_and_wait(node, 'bitdust set services/entangled-dht/udp-port "14441"')
-    # enable Stun server
-    run_ssh_command_and_wait(node, 'bitdust set services/ip-port-responder/enabled true')
-    # start BitDust daemon
-    start_daemon(node)
-    health_check(node)
-    print(f'\nSTARTED STUN SERVER [{node}]\n')
-
-
-def start_proxy_server(node, identity_name):
-    print(f'\nNEW PROXY SERVER {identity_name} at [{node}]\n')
-    # use short key to run tests faster
-    run_ssh_command_and_wait(node, 'bitdust set personal/private-key-size 1024')
-    # disable unwanted services
-    run_ssh_command_and_wait(node, 'bitdust set services/customer/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/supplier/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/proxy-transport/enabled false')
-
-    # configure ID servers
-    run_ssh_command_and_wait(node, 'bitdust set services/identity-propagate/min-servers 1')
-    run_ssh_command_and_wait(node, 'bitdust set services/identity-propagate/max-servers 1')
-    # configure DHT udp port
-    run_ssh_command_and_wait(node, 'bitdust set services/entangled-dht/udp-port "14441"')
-    # enable ProxyServer service
-    run_ssh_command_and_wait(node, 'bitdust set services/proxy-server/enabled true')
-    # start BitDust daemon and create new identity for proxy server
-    start_daemon(node)
-    health_check(node)
-    create_identity(node, identity_name)
-    connect_network(node)
-    print(f'\nSTARTED PROXY SERVER [{node}]\n')
-
-
-async def start_proxy_server_async(node, identity_name, loop):
-    print(f'\nNEW PROXY SERVER {identity_name} at [{node}]\n')
-    # use short key to run tests faster
-    await run_ssh_command_and_wait_async(node, 'bitdust set personal/private-key-size 1024', loop)
-    # disable unwanted services
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/customer/enabled false', loop)
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/supplier/enabled false', loop)
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/proxy-transport/enabled false', loop)
-
-    # configure ID servers
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/identity-propagate/min-servers 1', loop)
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/identity-propagate/max-servers 1', loop)
-    # configure DHT udp port
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/entangled-dht/udp-port "14441"', loop)
-    # enable ProxyServer service
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/proxy-server/enabled true', loop)
-    # start BitDust daemon and create new identity for proxy server
-    await start_daemon_async(node, loop)
-    await health_check_async(node, loop)
-    await create_identity_async(node, identity_name, loop)
-    await connect_network_async(node, loop)
-    print(f'\nSTARTED PROXY SERVER [{node}]\n')
-
-
-def start_supplier(node, identity_name):
-    print(f'\nNEW SUPPLIER {identity_name} at [{node}]\n')
-    # use short key to run tests faster
-    run_ssh_command_and_wait(node, 'bitdust set personal/private-key-size 1024')
-    # disable unwanted services
-    run_ssh_command_and_wait(node, 'bitdust set services/customer/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/proxy-server/enabled false')
-    # configure ID servers
-    run_ssh_command_and_wait(node, 'bitdust set services/identity-propagate/min-servers 1')
-    run_ssh_command_and_wait(node, 'bitdust set services/identity-propagate/max-servers 1')
-    # configure DHT udp port
-    run_ssh_command_and_wait(node, 'bitdust set services/entangled-dht/udp-port "14441"')
-    # set desired Proxy router
-    run_ssh_command_and_wait(node, f'bitdust set services/proxy-transport/preferred-routers "{PROXY_ROUTERS}"')
-    # enable supplier service
-    run_ssh_command_and_wait(node, 'bitdust set services/supplier/enabled true')
-    # start BitDust daemon and create new identity for supplier
-    start_daemon(node)
-    health_check(node)
-    create_identity(node, identity_name)
-    connect_network(node)
-    print(f'\nSTARTED SUPPLIER [{node}]\n')
-
-
-async def start_supplier_async(node, identity_name, loop):
-    print(f'\nNEW SUPPLIER {identity_name} at [{node}]\n')
-    # use short key to run tests faster
-    await run_ssh_command_and_wait_async(node, 'bitdust set personal/private-key-size 1024', loop)
-    # disable unwanted services
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/customer/enabled false', loop)
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/proxy-server/enabled false', loop)
-    # configure ID servers
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/identity-propagate/min-servers 1', loop)
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/identity-propagate/max-servers 1', loop)
-    # configure DHT udp port
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/entangled-dht/udp-port "14441"', loop)
-    # set desired Proxy router
-    await run_ssh_command_and_wait_async(node, f'bitdust set services/proxy-transport/preferred-routers "{PROXY_ROUTERS}"', loop)
-    # enable supplier service
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/supplier/enabled true', loop)
-    # start BitDust daemon and create new identity for supplier
-    await start_daemon_async(node, loop)
-    await health_check_async(node, loop)
-    await create_identity_async(node, identity_name, loop)
-    await connect_network_async(node, loop)
-    print(f'\nSTARTED SUPPLIER [{node}]\n')
-
-
-def start_customer(node, identity_name, join_network=True, num_suppliers=2):
-    print(f'\nNEW CUSTOMER {identity_name} at [{node}]\n')
-    # use short key to run tests faster
-    run_ssh_command_and_wait(node, 'bitdust set personal/private-key-size 1024')
-    # disable unwanted services
-    run_ssh_command_and_wait(node, 'bitdust set services/supplier/enabled false')
-    run_ssh_command_and_wait(node, 'bitdust set services/proxy-server/enabled false')
-    # configure ID servers
-    run_ssh_command_and_wait(node, 'bitdust set services/identity-propagate/min-servers 1')
-    run_ssh_command_and_wait(node, 'bitdust set services/identity-propagate/max-servers 1')
-    # configure DHT udp port
-    run_ssh_command_and_wait(node, 'bitdust set services/entangled-dht/udp-port "14441"')
-    # set desired Proxy router
-    run_ssh_command_and_wait(node, f'bitdust set services/proxy-transport/preferred-routers "{PROXY_ROUTERS}"')
-    # enable customer service and prepare tests
-    run_ssh_command_and_wait(node, 'bitdust set services/customer/enabled true')
-    run_ssh_command_and_wait(node, f'bitdust set services/customer/suppliers-number {num_suppliers}')
-    # create randomized file to test file upload/download
-    run_ssh_command_and_wait(node, f'python -c "import os, base64; print(base64.b64encode(os.urandom(24)).decode())" > /{node}/file_{node}.txt')
-    run_ssh_command_and_wait(node, f'python -c "import os, base64; print(base64.b64encode(os.urandom(24)).decode())" > /{node}/second_file_{node}.txt')
-    # start BitDust daemon and create new identity for supplier
-    start_daemon(node)
-    health_check(node)
-    if join_network:
-        create_identity(node, identity_name)
-        connect_network(node)
-    print(f'\nSTARTED CUSTOMER [{node}]\n')
-
-
-async def start_customer_async(node, identity_name, loop, join_network=True, num_suppliers=2):
-    print('\nNEW CUSTOMER %r at [%s]\n' % (identity_name, node, ))
-    # use short key to run tests faster
-    await run_ssh_command_and_wait_async(node, 'bitdust set personal/private-key-size 1024', loop)
-    # disable unwanted services
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/supplier/enabled false', loop)
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/proxy-server/enabled false', loop)
-    # configure ID servers
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/identity-propagate/min-servers 1', loop)
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/identity-propagate/max-servers 1', loop)
-    # configure DHT udp port
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/entangled-dht/udp-port "14441"', loop)
-    # set desired Proxy router
-    await run_ssh_command_and_wait_async(node, f'bitdust set services/proxy-transport/preferred-routers "{PROXY_ROUTERS}"', loop)
-    # enable customer service and prepare tests
-    await run_ssh_command_and_wait_async(node, 'bitdust set services/customer/enabled true', loop)
-    await run_ssh_command_and_wait_async(node, f'bitdust set services/customer/suppliers-number {num_suppliers}', loop)
-    # create randomized file to test file upload/download
-    await run_ssh_command_and_wait_async(node, f'python -c "import os, base64; print(base64.b64encode(os.urandom(24)).decode())" > /{node}/file_{node}.txt', loop)
-    await run_ssh_command_and_wait_async(node, f'python -c "import os, base64; print(base64.b64encode(os.urandom(24)).decode())" > /{node}/second_file_{node}.txt', loop)
-    # start BitDust daemon and create new identity for supplier
-    await start_daemon_async(node, loop)
-    await health_check_async(node, loop)
-    if join_network:
-        await create_identity_async(node, identity_name, loop)
-        await connect_network_async(node, loop)
-    print(f'\nSTARTED CUSTOMER [{node}]\n')
-
-
-#------------------------------------------------------------------------------
-
-async def open_one_tunnel_async(node, local_port, loop):
-    await open_tunnel_async(node, local_port, loop)
-
-#------------------------------------------------------------------------------
-
-def start_one_supplier(supplier):
-    start_supplier(node=supplier, identity_name=supplier)
-
-
-async def start_one_supplier_async(supplier, loop):
-    await start_supplier_async(node=supplier, identity_name=supplier, loop=loop)
-
-
-def start_one_customer(customer):
-    start_customer(
-        node=customer['name'],
-        identity_name=customer['name'],
-        join_network=customer['join_network'],
-        num_suppliers=customer['num_suppliers'],
-    )
-
-async def start_one_customer_async(customer, loop):
-    await start_customer_async(
-        node=customer['name'],
-        identity_name=customer['name'],
-        join_network=customer['join_network'],
-        num_suppliers=customer['num_suppliers'],
-        loop=loop,
-    )
-
-
-def start_one_proxy_server(proxy_server):
-    start_proxy_server(node=proxy_server, identity_name=proxy_server)
-
-
-async def start_one_proxy_server_async(proxy_server, loop):
-    await start_proxy_server_async(node=proxy_server, identity_name=proxy_server, loop=loop)
-
-#------------------------------------------------------------------------------
-
-def stop_one_node(node):
-    stop_daemon(node, skip_checks=True)
-
-#------------------------------------------------------------------------------
-
-def report_one_node(node):
-    main_log = run_ssh_command_and_wait(node, 'cat /root/.bitdust/logs/main.log')[0].strip()
-    num_warnings = main_log.count('WARNING')
-    num_errors = main_log.count('ERROR!!!')
-    num_exceptions = main_log.count('Exception:')
-    num_tracebacks = main_log.count('Traceback')
-    num_failures = main_log.count('Failure')
-    # assert num_exceptions == 0, 'found some critical errors in the log file on node %s' % node
-    print(f'[{node}]  Warnings: {num_warnings}     Errors: {num_errors}    Tracebacks: {num_tracebacks}     '
-          f'Failures: {num_failures}    Exceptions: {num_exceptions}')
-
-
-async def report_one_node_async(node, event_loop):
-    main_log = await run_ssh_command_and_wait_async(node, 'cat /root/.bitdust/logs/main.log', event_loop)
-    main_log = main_log[0].strip()
-    num_warnings = main_log.count('WARNING')
-    num_errors = main_log.count('ERROR!!!')
-    num_exceptions = main_log.count('Exception:')
-    num_tracebacks = main_log.count('Traceback')
-    num_failures = main_log.count('Failure')
-    # assert num_exceptions == 0, 'found some critical errors in the log file on node %s' % node
-    print(f'[{node}]  Warnings: {num_warnings}     Errors: {num_errors}    Tracebacks: {num_tracebacks}     '
-          f'Failures: {num_failures}    Exceptions: {num_exceptions}')
-
-
-def print_exceptions_one_node(node):
-    exceptions_out = run_ssh_command_and_wait(node, 'cat /root/.bitdust/logs/exception_*.log')[0].strip()
-    if exceptions_out:
-        print(f'\n[{node}]:\n\n{exceptions_out}\n\n')
-
-
-async def print_exceptions_one_node_async(node, event_loop):
-    exceptions_out = await run_ssh_command_and_wait_async(node, 'cat /root/.bitdust/logs/exception_*.log', event_loop)
-    exceptions_out = exceptions_out[0].strip()
-    if exceptions_out:
-        print(f'\n[{node}]:\n\n{exceptions_out}\n\n')
-
-#------------------------------------------------------------------------------
-
-async def clean_one_node_async(node, event_loop):
-    await run_ssh_command_and_wait_async(node, 'rm -rf /root/.bitdust/metadata', event_loop)
-    await run_ssh_command_and_wait_async(node, 'rm -rf /root/.bitdust/identitycache', event_loop)
-    await run_ssh_command_and_wait_async(node, 'rm -rf /root/.bitdust/identityserver', event_loop)
-    await run_ssh_command_and_wait_async(node, 'rm -rf /root/.bitdust/keys', event_loop)
-    await run_ssh_command_and_wait_async(node, 'rm -rf /root/.bitdust/customers', event_loop)
-    await run_ssh_command_and_wait_async(node, 'rm -rf /root/.bitdust/suppliers', event_loop)
-    await run_ssh_command_and_wait_async(node, 'rm -rf /root/.bitdust/backups', event_loop)
-    await run_ssh_command_and_wait_async(node, 'rm -rf /root/.bitdust/messages', event_loop)
-
-
-async def clean_one_customer_async(node, event_loop):
-    await run_ssh_command_and_wait_async(node, 'rm -rf /%s/*' % node, event_loop)
-
-
-def clean_one_node(node):
-    run_ssh_command_and_wait(node, 'rm -rf /root/.bitdust/metadata')
-    run_ssh_command_and_wait(node, 'rm -rf /root/.bitdust/identitycache')
-    run_ssh_command_and_wait(node, 'rm -rf /root/.bitdust/identityserver')
-    run_ssh_command_and_wait(node, 'rm -rf /root/.bitdust/keys')
-    run_ssh_command_and_wait(node, 'rm -rf /root/.bitdust/customers')
-    run_ssh_command_and_wait(node, 'rm -rf /root/.bitdust/suppliers')
-    run_ssh_command_and_wait(node, 'rm -rf /root/.bitdust/backups')
-    run_ssh_command_and_wait(node, 'rm -rf /root/.bitdust/messages')
-
-
-def clean_one_customer(node):
-    run_ssh_command_and_wait(node, 'rm -rf /%s/*' % node)
-
-#------------------------------------------------------------------------------
 
 def open_all_tunnels(event_loop):
+    _begin = time.time()
     print('\nStarting all SSH tunnels\n')
-    # event_loop.run_until_complete(asyncio.wait([
-    #     asyncio.ensure_future(open_one_tunnel_async(node, 10000+pos, event_loop)) for pos, node in enumerate(ALL_NODES)
-    # ]))
-    for node in ALL_NODES:
-        open_tunnel(node)
-    print('\nAll SSH tunnels opened\n')
+    event_loop.run_until_complete(asyncio.wait([
+        asyncio.ensure_future(ts.open_one_tunnel_async(node, 10000+pos, event_loop)) for pos, node in enumerate(ALL_NODES)
+    ]))
+    print('\nAll SSH tunnels opened in %5.3f seconds\n' % (time.time() - _begin))
 
 
 def clean_all_nodes(event_loop, skip_checks=False):
+    _begin = time.time()
     print('\nCleaning all nodes')
-    # event_loop.run_until_complete(asyncio.wait([
-    #     asyncio.ensure_future(clean_one_node_async(node, event_loop=event_loop)) for node in ALL_NODES
-    # ]))
-    for node in ALL_NODES:
-        clean_one_node(node)
-    # event_loop.run_until_complete(asyncio.wait([
-    #     asyncio.ensure_future(clean_one_customer_async(node['name'], event_loop)) for node in ALL_ROLES['customers']
-    # ]]))
-    for node in ALL_ROLES['customers']:
-        clean_one_customer(node['name'])
-    print('\n\nAll nodes cleaned')
+    event_loop.run_until_complete(asyncio.wait([
+        asyncio.ensure_future(ts.clean_one_node_async(node, event_loop=event_loop)) for node in ALL_NODES
+    ]))
+    event_loop.run_until_complete(asyncio.wait([
+        asyncio.ensure_future(ts.clean_one_customer_async(node['name'], event_loop=event_loop)) for node in ALL_ROLES['customers']
+    ]))
+
+    print('\n\nAll nodes cleaned in %5.3f seconds\n' % (time.time() - _begin))
 
 
 def start_all_nodes(event_loop):
+    _begin = time.time()
     print('\nStarting nodes\n')
 
     for number, dhtseed in enumerate(ALL_ROLES['dht-seeds']):
         # first seed to be started immediately, all other seeds must wait a bit before start
-        start_dht_seed(
+        ts.start_dht_seed(
             node=dhtseed['name'],
             other_seeds=dhtseed['other_seeds'],
             wait_seconds=(5 if number > 0 else 0),
@@ -655,74 +145,62 @@ def start_all_nodes(event_loop):
         )
     print('\nALL DHT SEEDS STARTED\n')
 
-    for idsrv in ALL_ROLES['identity-servers']:
-        start_identity_server(node=idsrv)
+    event_loop.run_until_complete(asyncio.wait([
+        ts.start_identity_server_async(idsrv, event_loop) for idsrv in ALL_ROLES['identity-servers']
+    ]))
     print('\nALL ID SERVERS STARTED\n')
 
-    for stunsrv in ALL_ROLES['stun-servers']:
-        start_stun_server(node=stunsrv)
+    event_loop.run_until_complete(asyncio.wait([
+        ts.start_stun_server_async(stunsrv, event_loop) for stunsrv in ALL_ROLES['stun-servers']
+    ]))
     print('\nALL STUN SERVERS STARTED\n')
 
-    for proxy_server in ALL_ROLES['proxy-servers']:
-        start_proxy_server(proxy_server, proxy_server)
-    # event_loop.run_until_complete(asyncio.wait([
-    #     start_one_proxy_server_async(proxy_server, event_loop) for proxy_server in ALL_ROLES['proxy-servers']
-    # ]))
+    event_loop.run_until_complete(asyncio.wait([
+        ts.start_one_proxy_server_async(proxy_server, event_loop) for proxy_server in ALL_ROLES['proxy-servers']
+    ]))
     print('\nALL PROXY SERVERS STARTED\n')
 
-    for supplier in ALL_ROLES['suppliers']:
-        start_supplier(supplier, supplier)
-    # event_loop.run_until_complete(asyncio.wait([
-    #     asyncio.ensure_future(start_one_supplier_async(supplier, event_loop)) for supplier in ALL_ROLES['suppliers']
-    # ]))
+    event_loop.run_until_complete(asyncio.wait([
+        asyncio.ensure_future(ts.start_one_supplier_async(supplier, event_loop)) for supplier in ALL_ROLES['suppliers']
+    ]))
     print('\nALL SUPPLIERS STARTED\n')
 
-    for customer in ALL_ROLES['customers']:
-        start_one_customer(customer)
-    # event_loop.run_until_complete(asyncio.wait([
-    #     asyncio.ensure_future(start_one_customer_async(customer, event_loop)) for customer in ALL_ROLES['customers']
-    # ]))
+    event_loop.run_until_complete(asyncio.wait([
+        asyncio.ensure_future(ts.start_one_customer_async(customer, event_loop)) for customer in ALL_ROLES['customers']
+    ]))
     print('\nALL CUSTOMERS STARTED\n')
 
-    print('\nDONE! ALL NODES STARTED!\n')
+    print('\nDONE! ALL NODES STARTED in %5.3f seconds\n' % (time.time() - _begin))
 
 
 def stop_all_nodes(event_loop):
+    _begin = time.time()
     print('\nstop all nodes\n')
-    # event_loop.run_until_complete(asyncio.wait([
-    #     asyncio.ensure_future(stop_one_node(node)) for node in ALL_NODES
-    # ]))
-    for node in ALL_NODES:
-        stop_one_node(node)
-    print('\nALL NODES STOPPED\n')
+    event_loop.run_until_complete(asyncio.wait([
+        asyncio.ensure_future(ts.stop_daemon_async(node, event_loop)) for node in ALL_NODES
+    ]))
+    print('\nALL NODES STOPPED in %5.3f seconds\n' % (time.time() - _begin))
 
 
 def kill_all_nodes():
     for node in ALL_NODES:
         print('Shutdown %s' % node)
-        run_ssh_command_and_wait(node, 'pkill -e sshd')
+        ts.run_ssh_command_and_wait(node, 'pkill -e sshd')
     print('All nodes stopped')
 
 
 def report_all_nodes(event_loop):
     print('\n\nTest report:')
-    # print('\n\nALL DHT RECORDS:')
-    # for node in ALL_NODES:
-    #     print('\n[%s]:' % node)
-    #     pprint.pprint(requests.get(tunnel_url(node, 'dht/db/dump/v1')).json())
+
     print('\n\nALL EXCEPTIONS:')
     for node in ALL_NODES:
-        print_exceptions_one_node(node)
-    # event_loop.run_until_complete(asyncio.wait([
-    #     asyncio.ensure_future(print_exceptions_one_node_async(node, event_loop)) for node in ALL_NODES
-    # ]))
-    # event_loop.run_until_complete(asyncio.wait([
-    #     asyncio.ensure_future(report_one_node_async(node, event_loop)) for node in ALL_NODES
-    # ]))
+        ts.print_exceptions_one_node(node)
+
     for node in ALL_NODES:
-        report_one_node(node)
+        ts.report_one_node(node)
 
 #------------------------------------------------------------------------------
+
 
 @pytest.yield_fixture(scope='session')
 def event_loop():
@@ -731,6 +209,7 @@ def event_loop():
     loop.close()
 
 #------------------------------------------------------------------------------
+
 
 @pytest.yield_fixture(scope='session', autouse=True)
 def global_wrapper(event_loop):
@@ -751,7 +230,7 @@ def global_wrapper(event_loop):
     if os.environ.get('START_NODES', '1') == '1':
         start_all_nodes(event_loop)
 
-    print('\nTest network prepared in %5.3f seconds\n')
+    print('\nTest network prepared in %5.3f seconds\n' % (time.time() - _begin))
  
     yield
